@@ -22,7 +22,7 @@
         "bin-dir": "bin"
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.0",
         "doctrine/annotations": "*",
         "symfony/finder": "*"
     },
@@ -35,7 +35,7 @@
     "require-dev": {
         "zendframework/zend-form": "<2.8",
         "squizlabs/php_codesniffer": ">=2.7",
-        "phpunit/phpunit": ">=4.8 <=5.6"
+        "phpunit/phpunit": ">=6.3"
     },
     "autoload-dev": {
         "psr-4": {
