@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -8,6 +8,7 @@ namespace Swagger\Annotations;
 
 use Exception;
 use Swagger\Analysis;
+use Swagger\Context;
 use Swagger\Logger;
 
 /**
@@ -19,91 +20,68 @@ use Swagger\Logger;
 class OpenApi extends AbstractAnnotation
 {
     /**
-     * Specifies the OpenAPI Specification version being used. It can be used by the Swagger UI and other clients to interpret the API listing.
+     * The semantic version number of the OpenAPI Specification version that the OpenAPI document uses.
+     * The openapi field should be used by tooling specifications and clients to interpret the OpenAPI document.
+     * This is not related to the API info.version string.
+     *
      * @var string
      */
     public $openapi = '3.0';
 
     /**
-     * Provides metadata about the API. The metadata can be used by the clients if needed.
+     * Provides metadata about the API. The metadata may be used by tooling as required.
+     *
      * @var Info
      */
     public $info;
 
     /**
-     * The host (name or ip) serving the API. This MUST be the host only and does not include the scheme nor sub-paths. It MAY include a port. If the host is not included, the host serving the documentation is to be used (including the port). The host does not support path templating.
-     * @var string
+     * An array of Server Objects, which provide connectivity information to a target server.
+     * If the servers property is not provided, or is an empty array, the default value would be a Server Object with a url value of /.
+     *
+     * @var Server[]
      */
-    public $host;
-
-    /**
-     * The base path on which the API is served, which is relative to the host. If it is not included, the API is served directly under the host. The value MUST start with a leading slash (/). The basePath does not support path templating.
-     * @var string
-     */
-    public $basePath;
-
-    /**
-     * The transfer protocol of the API. Values MUST be from the list: "http", "https", "ws", "wss". If the schemes is not included, the default scheme to be used is the one used to access the specification.
-     * @var array
-     */
-    public $schemes;
-
-    /**
-     * A list of MIME types the APIs can consume. This is global to all APIs but can be overridden on specific API calls. Value MUST be as described under Mime Types.
-     * @var array
-     */
-    public $consumes;
-
-    /**
-     * A list of MIME types the APIs can produce. This is global to all APIs but can be overridden on specific API calls. Value MUST be as described under Mime Types.
-     * @var array
-     */
-    public $produces;
+    public $servers;
 
     /**
      * The available paths and operations for the API.
+     *
      * @var Path[]
      */
     public $paths = [];
 
     /**
-     * An object to hold data types produced and consumed by operations.
-     * @var Definition[]
+     * An element to hold various components for the specification.
+     *
+     * @var Components
      */
-    public $definitions = [];
+    public $components;
+
 
     /**
-     * An object to hold parameters that can be used across operations. This property does not define global parameters for all operations.
-     * @var Parameter[]
-     */
-    public $parameters;
-
-    /**
-     * An object to hold responses that can be used across operations. This property does not define global responses for all operations.
-     * @var Response[]
-     */
-    public $responses;
-
-    /**
-     * Security scheme definitions that can be used across the specification.
-     * @var SecurityScheme[]
-     */
-    public $securityDefinitions;
-
-    /**
-     * A declaration of which security schemes are applied for the API as a whole. The list of values describes alternative security schemes that can be used (that is, there is a logical OR between the security requirements). Individual operations can override this definition.
-     * @var array
+     * A declaration of which security mechanisms can be used across the API.
+     * The list of values includes alternative security requirement objects that can be used.
+     * Only one of the security requirement objects need to be satisfied to authorize a request.
+     * Individual operations can override this definition.
+     *
+     * @var SecurityRequirement[]
      */
     public $security;
 
     /**
-     * A list of tags used by the specification with additional metadata. The order of the tags can be used to reflect on their order by the parsing tools. Not all tags that are used by the Operation Object must be declared. The tags that are not declared may be organized randomly or based on the tools' logic. Each tag name in the list MUST be unique.
+     * A list of tags used by the specification with additional metadata.
+     * The order of the tags can be used to reflect on their order by the parsing tools.
+     * Not all tags that are used by the Operation Object must be declared.
+     * The tags that are not declared may be organized randomly or based on the tools' logic.
+     * Each tag name in the list must be unique.
+     *
      * @var Tag[]
      */
     public $tags;
 
     /**
      * Additional external documentation.
+     *
      * @var ExternalDocumentation
      */
     public $externalDocs;
@@ -122,23 +100,24 @@ class OpenApi extends AbstractAnnotation
     /** @inheritdoc */
     public static $_nested = [
         'Swagger\Annotations\Info' => 'info',
+        'Swagger\Annotations\Server' => ['servers', 'server'],
         'Swagger\Annotations\Path' => ['paths', 'path'],
-        'Swagger\Annotations\Definition' => ['definitions', 'definition'],
+        'Swagger\Annotations\Components' => 'components',
         'Swagger\Annotations\Tag' => ['tags'],
-        'Swagger\Annotations\Parameter' => ['parameters', 'parameter'],
-        'Swagger\Annotations\Response' => ['responses', 'response'],
         'Swagger\Annotations\ExternalDocumentation' => 'externalDocs',
-        'Swagger\Annotations\SecurityScheme' => ['securityDefinitions', 'securityDefinition']
+        'Swagger\Annotations\SecurityRequirement' => ['security', 'securityRequirement']
     ];
 
     /** @inheritdoc */
-    public static $_types = [
-        'host' => 'string',
-        'basePath' => 'string',
-        'schemes' => '[scheme]',
-        'consumes' => '[string]',
-        'produces' => '[string]',
-    ];
+    public static $_types = [];
+
+    public function __construct($properties)
+    {
+        parent::__construct($properties);
+        if ($this->components === null) {
+            $this->components = new Components(['_context'=> new Context(['nested' => $this], $this->_context)]);
+        }
+    }
 
     /** @inheritdoc */
     public function validate($parents = null, $skip = null, $ref = null)
