@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -66,6 +66,7 @@ class Response extends AbstractAnnotation
 
     /** @inheritdoc */
     public static $_parents = [
+        'Swagger\Annotations\Components',
         'Swagger\Annotations\Operation',
         'Swagger\Annotations\Get',
         'Swagger\Annotations\Post',
@@ -74,6 +75,5 @@ class Response extends AbstractAnnotation
         'Swagger\Annotations\Delete',
         'Swagger\Annotations\Head',
         'Swagger\Annotations\Options',
-        'Swagger\Annotations\OpenApi'
     ];
 }
