@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -11,7 +11,7 @@ class AugmentParameterTest extends SwaggerTestCase
     public function testAugmentParameter()
     {
         $openapi = \Swagger\scan(__DIR__ . '/Fixtures/UsingRefs.php');
-        $this->assertCount(1, $openapi->parameters, 'Swagger contains 1 reusable parameter specification');
-        $this->assertEquals('ItemName', $openapi->parameters[0]->parameter, 'When no @SWG\Parameter()->parameter is specified, use @SWG\Parameter()->name');
+        $this->assertCount(1, $openapi->components->parameters, 'Swagger contains 1 reusable parameter specification');
+        $this->assertEquals('ItemName', $openapi->components->parameters[0]->parameter, 'When no @SWG\Parameter()->parameter is specified, use @SWG\Parameter()->name');
     }
 }
