@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -6,8 +6,9 @@
 
 namespace SwaggerTests;
 
-use Swagger\Processors\AugmentDefinitions;
+use Swagger\Processors\AugmentSchemas;
 use Swagger\Processors\AugmentProperties;
+use Swagger\Processors\MergeIntoComponents;
 use Swagger\Processors\MergeIntoOpenApi;
 use Swagger\StaticAnalyser;
 
@@ -18,29 +19,53 @@ class AugmentPropertiesTest extends SwaggerTestCase
         $analyser = new StaticAnalyser();
         $analysis = $analyser->fromFile(__DIR__ . '/Fixtures/Customer.php');
         $analysis->process(new MergeIntoOpenApi());
-        $this->assertCount(1, $analysis->openapi->definitions);
-        $customer = $analysis->openapi->definitions[0];
-        $this->assertSame(null, $customer->properties, 'Sanity check. @SWG\Property\'s not yet erged ');
-        $analysis->process(new AugmentDefinitions());
-        $analysis->process(new AugmentProperties());
+        $analysis->process(new MergeIntoComponents());
+        $analysis->process(new AugmentSchemas());
+        $customer = $analysis->openapi->components->schemas[0];
         $firstname = $customer->properties[0];
+        $lastname = $customer->properties[1];
+        $tags = $customer->properties[2];
+        $submittedBy = $customer->properties[3];
+        $friends = $customer->properties[4];
+
+        // Verify no values where defined in the annotation.
+        $this->assertNull($firstname->property);
+        $this->assertNull($firstname->description);
+        $this->assertNull($firstname->type);
+
+        $this->assertNull($lastname->property);
+        $this->assertNull($lastname->description);
+        $this->assertNull($lastname->type);
+
+        $this->assertNull($tags->property);
+        $this->assertNull($tags->type);
+        $this->assertNull($tags->items);
+
+        $this->assertNull($submittedBy->property);
+        $this->assertNull($submittedBy->ref);
+
+        $this->assertNull($friends->property);
+        $this->assertNull($friends->type);
+
+        $analysis->process(new AugmentProperties());
+
         $this->assertSame('firstname', $firstname->property, '@SWG\Property()->property based on propertyname');
         $this->assertSame('The firstname of the customer.', $firstname->description, '@SWG\Property()->description based on @var description');
         $this->assertSame('string', $firstname->type, '@SWG\Property()->type based on @var declaration');
-        $lastname = $customer->properties[1];
+
         $this->assertSame('lastname', $lastname->property);
         $this->assertSame('The lastname of the customer.', $lastname->description);
         $this->assertSame('string', $lastname->type);
-        $tags = $customer->properties[2];
+
         $this->assertSame('tags', $tags->property);
         $this->assertSame('array', $tags->type, 'Detect array notation: @var string[]');
         $this->assertSame('string', $tags->items->type);
-        $submittedBy = $customer->properties[3];
+
         $this->assertSame('submittedBy', $submittedBy->property);
-        $this->assertSame('#/definitions/Customer', $submittedBy->ref);
-        $friends = $customer->properties[4];
+        $this->assertSame('#/components/schemas/Customer', $submittedBy->ref);
+
         $this->assertSame('friends', $friends->property);
         $this->assertSame('array', $friends->type);
-        $this->assertSame('#/definitions/Customer', $friends->items->ref);
+        $this->assertSame('#/components/schemas/Customer', $friends->items->ref);
     }
 }
