@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -31,9 +31,9 @@ END;
         $analysis = new Analysis($this->parseComment($comment));
         $this->assertCount(3, $analysis->annotations);
         $analysis->process(new MergeIntoOpenApi());
-        $this->assertCount(4, $analysis->annotations);
+        $this->assertCount(5, $analysis->annotations);
         $before = $analysis->split();
-        $this->assertCount(2, $before->merged->annotations, 'Generated @SWG\OpenApi + @SWG\Info');
+        $this->assertCount(3, $before->merged->annotations, 'Generated @SWG\OpenApi, @SWG\Components and @SWG\Info');
         $this->assertCount(2, $before->unmerged->annotations, '@SWG\License + @SWG\Contact');
         $this->assertCount(0, $analysis->openapi->_unmerged);
         $analysis->validate(); // Validation fails to detect the unmerged annotations.
@@ -41,7 +41,7 @@ END;
         // CleanUnmerged should place the unmerged annotions into the swagger->_unmerged array.
         $analysis->process(new CleanUnmerged());
         $between = $analysis->split();
-        $this->assertCount(2, $between->merged->annotations, 'Generated @SWG\OpenApi + @SWG\Info');
+        $this->assertCount(3, $between->merged->annotations, 'Generated @SWG\OpenApi, @SWG\Components and @SWG\Info');
         $this->assertCount(2, $between->unmerged->annotations, '@SWG\License + @SWG\Contact');
         $this->assertCount(2, $analysis->openapi->_unmerged); // 1 would also be oke, Could a'Only the @SWG\License'
         $this->assertSwaggerLogEntryStartsWith('Unexpected @SWG\License(), expected to be inside @SWG\Info in ');
@@ -56,7 +56,7 @@ END;
         $analysis->process(new CleanUnmerged());
         $this->assertCount(0, $license->_unmerged);
         $after = $analysis->split();
-        $this->assertCount(3, $after->merged->annotations, 'Generated @SWG\OpenApi + @SWG\Info + @SWG\Contact');
+        $this->assertCount(4, $after->merged->annotations, 'Generated @SWG\OpenApi, @SWG\Components, @SWG\Info and @SWG\Contact');
         $this->assertCount(1, $after->unmerged->annotations, '@SWG\License');
         $this->assertCount(1, $analysis->openapi->_unmerged);
         $this->assertSwaggerLogEntryStartsWith('Unexpected @SWG\License(), expected to be inside @SWG\Info in ');
