@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @license Apache 2.0
@@ -6,8 +6,9 @@
 
 namespace SwaggerTests;
 
-use Swagger\Processors\AugmentDefinitions;
+use Swagger\Processors\MergeIntoComponents;
 use Swagger\Processors\AugmentProperties;
+use Swagger\Processors\AugmentSchemas;
 use Swagger\Processors\MergeIntoOpenApi;
 use Swagger\StaticAnalyser;
 
@@ -18,15 +19,16 @@ class NestedPropertyTest extends SwaggerTestCase
         $analyser = new StaticAnalyser();
         $analysis = $analyser->fromFile(__DIR__ . '/Fixtures/NestedProperty.php');
         $analysis->process(new MergeIntoOpenApi());
-        $analysis->process(new AugmentDefinitions());
+        $analysis->process(new MergeIntoComponents());
+        $analysis->process(new AugmentSchemas());
         $analysis->process(new AugmentProperties());
-        
-        $this->assertCount(1, $analysis->openapi->definitions);
-        $definition = $analysis->openapi->definitions[0];
-        $this->assertEquals('NestedProperty', $definition->definition);
-        $this->assertCount(1, $definition->properties);
 
-        $parentProperty = $definition->properties[0];
+        $this->assertCount(1, $analysis->openapi->components->schemas);
+        $schema = $analysis->openapi->components->schemas[0];
+        $this->assertEquals('NestedProperty', $schema->schema);
+        $this->assertCount(1, $schema->properties);
+
+        $parentProperty = $schema->properties[0];
         $this->assertEquals('parentProperty', $parentProperty->property);
         $this->assertCount(1, $parentProperty->properties);
 
@@ -37,7 +39,7 @@ class NestedPropertyTest extends SwaggerTestCase
         $theBabyOfBaby = $babyProperty->properties[0];
         $this->assertEquals('theBabyOfBaby', $theBabyOfBaby->property);
         $this->assertCount(1, $theBabyOfBaby->properties);
-     
+
         // verbose not-recommend notations
         $theBabyOfBabyBaby = $theBabyOfBaby->properties[0];
         $this->assertEquals('theBabyOfBabyBaby', $theBabyOfBabyBaby->property);
